+++ events/message/lvl.js
@@ -7,6 +7,8 @@ exports.run = async (message, client) =>{
     if(message.content.includes(prefix)) return
     if(message.member.roles.filter(role => role.name === "muted").size !==0) return
     if (fs.existsSync("./config/guild/" + message.guild.id + "/lvl/" + message.author.id + " lvl.json")){
+        if (!fs.existsSync("./config/guild/" + message.guild.id + "/lvl/enabled")) return fs.writeFileSync("./config/guild/" + message.guild.id + "/lvl/enabled", "off")
+        if (!fs.existsSync("./config/guild/" + message.guild.id + "/lvl/enabled") || !(fs.readFileSync("./config/guild/" + message.guild.id + "/lvl/enabled", "utf-8") === "on")) return
         log.log("[JeuxGate : elvl] Created lvl ", "messages", message.guild.id + "/lvl")
         if(!fs.existsSync("./config/guild/" + message.guild.id + "/lvl/" + message.author.id + " cooldown")){
             fs.writeFileSync("./config/guild/" + message.guild.id + "/lvl/" + message.author.id + " cooldown", "on")
@@ -30,9 +32,11 @@ exports.run = async (message, client) =>{
     }else{
         log.log("[JeuxGate : elvl] Creating lvl ", "messages", message.guild.id + "/lvl")
         if(fs.existsSync("./config/guild/" + message.guild.id + "/lvl/")){
+            fs.writeFileSync("./config/guild/" + message.guild.id + "/lvl/enabled", "off")
             fs.writeFileSync("./config/guild/" + message.guild.id + "/lvl/" + message.author.id + " lvl.json", "{\"level\": 0,\"messagesend\": 0}")
         }else{
             fs.mkdirSync("./config/guild/" + message.guild.id + "/lvl/")
+            fs.writeFileSync("./config/guild/" + message.guild.id + "/lvl/enabled", "off")
             fs.writeFileSync("./config/guild/" + message.guild.id + "/lvl/" + message.author.id + " lvl.json", "{\"level\": 0,\"messagesend\": 0}")
         }
     }
